+++ applications/TrilinosApplication/tests/cpp_tests/test_builder_and_solver.cpp
@@ -22,7 +22,7 @@
 // Project includes
 #include "testing/testing.h"
 #include "spaces/ublas_space.h"
-#include "custom_utilities/trilinos_cpp_test_utilities.h"
+#include "trilinos_cpp_test_utilities.h"
 #include "containers/model.h"
 #include "mpi/includes/mpi_data_communicator.h"
 #include "mpi/utilities/parallel_fill_communicator.h"
