+++ applications/TrilinosApplication/tests/cpp_tests/test_cpp_test_utililies.cpp
@@ -17,7 +17,7 @@
 // Project includes
 #include "testing/testing.h"
 #include "mpi/includes/mpi_data_communicator.h"
-#include "custom_utilities/trilinos_cpp_test_utilities.h"
+#include "trilinos_cpp_test_utilities.h"
 
 namespace Kratos::Testing
 {
